Re-write BP7 tuning data to conform to Tuning.xsd.
BP7-tuning2.xml contains the original tuning data.
BP7 contains the same data in the Tuning.xsd schema.
Revised Tuning.xsd to allow <frequency> to be optional--there is one fingering/register
in BP7 that has no target frequency.  At some point, we'll need to add something to
the schema to allow us to tell the modelling algorithms which playing range this note applies to.