Introduce ObjectiveFunction and Evaluator classes on which to perform optimization. An Evaluator encapsulates an instrument calculator and uses it to perform a specific evaluation on instrument tuning at specified fingerings.
An ObjectiveFunction implements a MultivariateFunction or UnivariateFunction,
and everything else we might need to feed an Apache Commons Optimizer.
Each ObjectiveFunction class manages specific aspects of an instrument geometry
that we want to vary in an optimization.  When instantiated with a calculator
and an evaluator, it can vary the instrument geometry, and evaluate the outcome.